# Conflicts:
#	Elsa.sln
#	src/modules/Elsa.EntityFrameworkCore/Modules/Management/WorkflowDefinitionStore.cs

+++ src/modules/Elsa.ProtoActor/Services/ProtoActorWorkflowRuntime.cs
@@ -252,7 +252,7 @@ public class ProtoActorWorkflowRuntime : IWorkflowRuntime
     }
 
     /// <inheritdoc />
-    public async Task<int> CountRunningWorkflowsAsync(CountRunningWorkflowsArgs args, CancellationToken cancellationToken = default)
+    public async Task<long> CountRunningWorkflowsAsync(CountRunningWorkflowsArgs args, CancellationToken cancellationToken = default)
     {
         var client = _cluster.GetNamedRunningWorkflowsGrain();
